Extract unit test for BulkImportStarterLambda
