+++ java/bulk-import/bulk-import-starter/src/test/java/sleeper/bulkimport/starter/BulkImportStarterLambdaTestHelper.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022-2023 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.bulkimport.starter;
+
+import com.amazonaws.services.lambda.runtime.events.SQSEvent;
+import com.google.common.collect.Lists;
+
+public class BulkImportStarterLambdaTestHelper {
+    private BulkImportStarterLambdaTestHelper() {
+    }
+
+    public static SQSEvent getSqsEvent(String message) {
+        SQSEvent event = new SQSEvent();
+        SQSEvent.SQSMessage sqsMessage = new SQSEvent.SQSMessage();
+        sqsMessage.setBody(message);
+        event.setRecords(Lists.newArrayList(
+                sqsMessage
+        ));
+        return event;
+    }
+}
